requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1

# AI/ML Dependencies
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.0.0
tensorflow>=2.15.0
sentencepiece>=0.1.99
protobuf>=3.20.0,<4.0

# NLP
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
openai-whisper>=20230918
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Caching
redis>=5.0.0
diskcache>=5.6.0

# API/Web
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.4.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
python-dateutil>=2.8.2
jsonschema>=4.19.0
pyyaml>=6.0
click>=8.1.0

# File Processing
python-magic>=0.4.27
pillow>=10.0.0

# Monitoring/Logging
loguru>=0.7.0
sentry-sdk>=1.35.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Export Functionality
reportlab>=4.0.0
openpyxl>=3.1.0